(notepadplusplus.install) Add \NoStop install parameter to avoid killing notepad++ processes

+++ automatic/notepadplusplus.install/tools/chocolateyBeforeModify.ps1
@@ -1,10 +1,17 @@
-﻿$process = Get-Process "Notepad++*" -ea 0
+﻿$pp = Get-PackageParameters
+
+$process = Get-Process "Notepad++*" -ea 0
 
 if ($process) {
   $runningFile = "$env:TEMP\npp.running"
   $processPath = $process | Where-Object { $_.Path } | Select-Object -First 1 -ExpandProperty Path
   Set-Content -Value "$processPath" -Path $runningFile
 
-  Write-Host "Found Running instance of Notepad++. Stopping processes..."
-  $process | Stop-Process
+  if ($pp.NoStop) {
+    Write-Warning "Not stopping running Notepad++ process"  
+  }
+  else {
+    Write-Host "Found Running instance of Notepad++. Stopping processes..."
+    $process | Stop-Process
+  }
 }
